Fix: SignIn and SignUp pages for mobile
