+++ src/App.js
@@ -1,9 +1,9 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { GlobalStyle } from "./globalStyles.js";
 import AuthProvider from "./contexts/auth.context.js";
-import SignUp from "../src/components/SignUp.js";
-import SignIn from "./components/SignIn.js";
-//Quem estiver fazendo a rota timeline, por favor remove o comentário no components SingIn quando recebe a resposta da promise por favor.
+import SignUp from "../src/components/AuthPage/SignUp.js";
+import SignIn from "../src/components/AuthPage/SignIn.js";
+
 function App() {
   return (<>
     <GlobalStyle />
